src/app/app.component.ts
import { Component, ViewChild } from '@angular/core';
import { Router } from '@angular/router';
import { MatSidenav } from '@angular/material/sidenav';
import { BreakpointObserver } from '@angular/cdk/layout';

@Component({
  selector: 'app-root',
  templateUrl: './app.component.html',
  styleUrls: ['./app.component.css']
})
export class AppComponent {
  title = '2022國立臺海大就業博覽會';

  @ViewChild(MatSidenav)
  sidenav!: MatSidenav;

  public constructor(private router: Router, private observer: BreakpointObserver) {

  }

  public negativePage(pageName: string){
    this.router.navigate([pageName]);
  }

  ngAfterViewInit() {
    this.observer.observe(['(max-width: 800px)']).subscribe((res) => {
      if(res.matches){
        this.sidenav.mode = 'over';
        this.sidenav.close();
      }
      else{
        this.sidenav.mode = 'side';
        this.sidenav.open();
      }
    });
  }
}
